move preflight requirements check to BaseConfig
