+++ reV/config/offshore_config.py
@@ -6,7 +6,6 @@ reV offshore wind aggregation config.
 """
 import logging
 
-from reV.utilities.exceptions import ConfigError
 from reV.config.base_analysis_config import AnalysisConfig
 from reV.pipeline.pipeline import Pipeline
 
@@ -29,19 +28,6 @@ class OffshoreConfig(AnalysisConfig):
             or dictionary with pre-extracted config.
         """
         super().__init__(config)
-        self._offshore_preflight()
-
-    def _offshore_preflight(self):
-        """Perform pre-flight checks on the SC agg config inputs"""
-        missing = []
-        for req in self.REQUIREMENTS:
-            if self.get(req, None) is None:
-                missing.append(req)
-        if any(missing):
-            e = ('SC offshore config missing the following '
-                 'keys: {}'.format(missing))
-            logger.error(e)
-            raise ConfigError(e)
 
     @property
     def gen_fpath(self):
