+++ reV/config/supply_curve_configs.py
@@ -48,16 +48,9 @@ class SupplyCurveAggregationConfig(AnalysisConfig):
 
     def _sc_agg_preflight(self):
         """Perform pre-flight checks on the SC agg config inputs"""
-        missing = []
-        for req in self.REQUIREMENTS:
-            if self.get(req, None) is None:
-                missing.append(req)
-        if any(missing):
-            raise ConfigError('SC Aggregation config missing the following '
-                              'keys: {}'.format(missing))
-
         with h5py.File(self.excl_fpath, mode='r') as f:
             dsets = list(f)
+
         if self.tm_dset not in dsets and self.res_fpath is None:
             raise ConfigError('Techmap dataset "{}" not found in exclusions '
                               'file, resource file input "res_fpath" is '
